project_news/brain/requirements.txt
python-dotenv==1.0.0
anyio==3.7.0
APScheduler==3.10.1
blis==0.7.9
boto3==1.26.147
botocore==1.29.147
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
DAWG-Python==0.7.2
docopt==0.6.2
dostoevsky==0.6.0
exceptiongroup==1.1.2
fastapi==0.99.1
fasttext==0.9.2
fonttools==4.39.4
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.7.1
murmurhash==1.0.9
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
psycopg2==2.9.6
pybind11==2.10.4
pydantic==1.10.9
pymorphy3==1.2.0
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
razdel==0.5.0
reportlab==4.0.4
requests==2.31.0
ru-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_lg-3.5.0/ru_core_news_lg-3.5.0-py3-none-any.whl
s3transfer==0.6.1
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.16
srsly==2.4.6
starlette==0.27.0
thinc==8.1.10
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
uvicorn==0.22.0
wasabi==1.1.2
zipp==3.15.0
